requirements.txt
asgiref==3.11.0
Django==6.0.1
pillow==12.1.0
sqlparse==0.5.5
python-magic==0.4.27
dj-database-url==2.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
easy-thumbnails==2.10.0

